# openapi.json
{
  "components": {
    "schemas": {
...
      "AggregateReport": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/AggregateRow"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/AggregateMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "AggregateReport",
        "type": "object"
      },
...
      "Bucket": {
        "enum": [
          "day",
          "week",
          "month"
        ],
        "title": "Bucket",
        "type": "string"
      },
...
      "CostPerUnitReport": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/CostPerUnitRow"
            },
            "title": "Data",
            "type": "array"
          },
          "unit": {
            "$ref": "#/components/schemas/Unit"
          }
        },
        "required": [
          "data",
          "unit"
        ],
        "title": "CostPerUnitReport",
        "type": "object"
      },
...
      "EventType": {
        "enum": [
          "production",
          "expense",
          "income",
          "observation",
          "reproductive",
          "acquisition",
          "mortality",
          "inventory"
        ],
        "title": "EventType",
        "type": "string"
      },
...
      "GroupBy": {
        "enum": [
          "asset"
        ],
        "title": "GroupBy",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InventoryAdjustment": {
        "enum": [
          "increment",
          "decrement",
          "reset"
        ],
        "title": "InventoryAdjustment",
        "type": "string"
      },
...
      "Unit": {
        "enum": [
          "g",
          "kg",
          "lb",
          "t",
          "ml",
          "l",
          "gal",
          "unit",
          "dozen",
          "head"
        ],
        "title": "Unit",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Ovejitas API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/farms/{farm_id}/reports/aggregate": {
      "get": {
        "description": "Dispatches on `type` and returns uniform `{bucket, group, group_label, measure, value, asset_id}` rows.\n\n- production / observation: SUM(quantity) grouped by unit\n- mortality / acquisition: SUM(quantity) as headcount, no grouping\n- inventory: net flow within window (increments minus decrements).   Pass `adjustment=reset|increment|decrement` to isolate one kind.\n- expense / income: SUM(amount) grouped by currency\n- reproductive: COUNT(*) of events\n\nFilters `unit`, `adjustment`, `currency` are ignored for types where they do not apply.\n\n**`group_by=asset`** breaks rows down per asset. Each row then carries a stable `group` key (the asset id as a string), a `group_label` (the asset name), and an `asset_id`. When `group_by` is omitted, rows are unchanged: `group_label` and `asset_id` stay `null`.\n\nCompatibility matrix \u2014 `type` vs `group_by`:\n\n| type | group_by=asset |\n| --- | --- |\n| mortality | supported |\n| acquisition | supported |\n| production | supported (quantity summed per asset across units) |\n| observation / inventory / expense / income / reproductive | rejected with 422 |",
        "operationId": "reports_aggregate_report",
        "parameters": [
          {
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "title": "Farm Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EventType"
            }
          },
          {
            "in": "query",
            "name": "bucket",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Bucket",
              "default": "day"
            }
          },
          {
            "in": "query",
            "name": "group_by",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/GroupBy"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Group By"
            }
          },
          {
            "in": "query",
            "name": "asset_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset Id"
            }
          },
          {
            "in": "query",
            "name": "unit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Unit"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Unit"
            }
          },
          {
            "in": "query",
            "name": "adjustment",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/InventoryAdjustment"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Adjustment"
            }
          },
          {
            "in": "query",
            "name": "currency",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Currency"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AggregateReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Generic time-bucketed aggregate over events of one type",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/farms/{farm_id}/reports/cost-per-unit": {
      "get": {
        "description": "Requires `unit` (what counts as one produced unit). One row per producer asset (any asset with `production` events in that unit). `cost_per_unit = (direct expense events on the producer + the average-cost value of the feed it was fed) / its production quantity`. Feed is attributed via `material_consumption` with `reason=feeding` and `consumer_asset_id` = the producer; a material's average cost is its full purchase history (not bounded by `date_from`). `date_from`/`date_to` bound production and direct expenses. A producer that made nothing in the window still appears with `cost_per_unit` null; `has_unvalued_consumption` flags rows whose feed has no purchase history to value it.",
        "operationId": "reports_cost_per_unit",
        "parameters": [
          {
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "title": "Farm Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "in": "query",
            "name": "asset_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset Id"
            }
          },
          {
            "in": "query",
            "name": "unit",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Unit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CostPerUnitReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "R3 \u2014 cost per produced unit, per producer asset",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/farms/{farm_id}/reports/cost-per-unit/pdf": {
      "get": {
        "operationId": "reports_cost_per_unit_pdf",
        "parameters": [
          {
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "title": "Farm Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "in": "query",
            "name": "asset_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset Id"
            }
          },
          {
            "in": "query",
            "name": "unit",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Unit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "R3 \u2014 PDF download",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/farms/{farm_id}/reports/individuals/{individual_id}/timeline": {
      "get": {
        "operationId": "reports_timeline",
        "parameters": [
          {
            "in": "path",
            "name": "individual_id",
            "required": true,
            "schema": {
              "title": "Individual Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "title": "Farm Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EventType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_EventRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "R4 \u2014 paginated event timeline for one individual",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/farms/{farm_id}/reports/inventory-summary": {
      "get": {
        "description": "One row per (asset, unit) for any asset that carries INVENTORY events \u2014 material assets and aggregated animal flocks. On-hand is derived from those events: sum of increments minus decrements since the most recent reset. `date_to` gives the balance as of that moment (default: now); `date_from` does not apply to a running balance and is ignored.",
        "operationId": "reports_inventory_summary",
        "parameters": [
          {
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "title": "Farm Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "in": "query",
            "name": "asset_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventorySummaryReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "R5 \u2014 current on-hand inventory per asset",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/farms/{farm_id}/reports/material-consumption-aggregate": {
      "get": {
        "description": "Time-bucketed SUM(quantity) over recorded material consumptions.\n\n- `bucket=day|week|month` \u2014 `date_trunc` window\n- `group_by=material|consumer|both` \u2014 each row carries a `group` key, a `group_label`, and the `unit` (quantities never sum across units)\n- optional filters: `material_asset_id`, `consumer_asset_id`, `reason`, `date_from`, `date_to`\n\n`totals` carries the per-(group, unit) `total_qty` across all buckets.",
        "operationId": "reports_material_consumption_aggregate",
        "parameters": [
          {
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "title": "Farm Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "in": "query",
            "name": "bucket",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Bucket",
              "default": "day"
            }
          },
          {
            "in": "query",
            "name": "group_by",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ConsumptionGroupBy",
              "default": "material"
            }
          },
          {
            "in": "query",
            "name": "material_asset_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Material Asset Id"
            }
          },
          {
            "in": "query",
            "name": "consumer_asset_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Consumer Asset Id"
            }
          },
          {
            "in": "query",
            "name": "reason",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ConsumptionReason"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reason"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialConsumptionAggregateReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Day/week material-consumption totals, bucketed and grouped",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/farms/{farm_id}/reports/profitability": {
      "get": {
        "description": "One row per (asset, currency). Events with NULL amount/currency are excluded. Different currencies are never silently summed.",
        "operationId": "reports_profitability",
        "parameters": [
          {
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "title": "Farm Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "in": "query",
            "name": "asset_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfitabilityReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "R1 \u2014 income minus expense per asset",
        "tags": [
          "reports"
        ]
      }
    },
...
  }
}